requirements.txt
pika
python-dotenv
psycopg2
requests
pydub
python-telegram-bot
Pillow
jdatetime